+++ endianness.h
@@ -1,18 +1,20 @@
 #ifndef ENDIANNESS_H
 #define ENDIANNESS_H
 
+#include <stdint.h>
+
 #define L_ENDIAN 0
 #define B_ENDIAN    1
 
 int get_endianness();
 
-unsigned long long swpd(double d);
-double uswpd(unsigned long long d);
+uint64_t swpd(double d);
+double uswpd(uint64_t d);
 
 float swapf(float);
 double swapd(double);
-void swaps(unsigned short *x);
-void swapi(unsigned int *x);
-void swapl(unsigned long *x);
+void swaps(uint16_t *x);
+void swapi(uint32_t *x);
+void swapl(uint64_t *x);
 
 #endif
